Move game detection and communication logic to seperate services
