+++ H2MLauncher.Core/Services/H2MCommunicationService.cs
@@ -1,11 +1,10 @@
 ﻿using System.Data;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.Net;
 using System.Runtime.InteropServices;
 using System.Text;
 
 using H2MLauncher.Core.Settings;
+using H2MLauncher.Core.Utilities;
 
 using Microsoft.Extensions.Logging;
 
@@ -16,7 +15,6 @@ namespace H2MLauncher.Core.Services
     public sealed class H2MCommunicationService : IDisposable
     {
         private const string GAME_WINDOW_TITLE = "H2M-Mod";
-        private const string GAME_EXECUTABLE_NAME = "h1_mp64_ship.exe";
 
         //Windows API constants
         private const int WM_CHAR = 0x0102; // Message code for sending a character
@@ -28,368 +26,48 @@ namespace H2MLauncher.Core.Services
         private readonly ILogger<H2MCommunicationService> _logger;
         private readonly IDisposable? _optionsChangeRegistration;
 
+        private readonly IGameCommunicationService _gameCommunicationService;
+        private readonly IGameDetectionService _gameDetectionService;
+
         public H2MCommunicationService(IErrorHandlingService errorHandlingService, IWritableOptions<H2MLauncherSettings> options,
-            ILogger<H2MCommunicationService> logger)
+            ILogger<H2MCommunicationService> logger, IGameCommunicationService gameCommunicationService, IGameDetectionService gameDetectionService)
         {
             _errorHandlingService = errorHandlingService;
             _h2mLauncherSettings = options;
             _logger = logger;
+            _gameCommunicationService = gameCommunicationService;
+            _gameDetectionService = gameDetectionService;
 
             if (options.CurrentValue.AutomaticGameDetection)
             {
-                StartGameDetection();
+                _gameDetectionService.StartGameDetection();
             }
 
             _optionsChangeRegistration = options.OnChange((settings, _) =>
             {
-                if (!settings.AutomaticGameDetection && IsGameDetectionRunning)
+                if (!settings.AutomaticGameDetection && _gameDetectionService.IsGameDetectionRunning)
                 {
-                    StopGameDetection();
+                    _gameDetectionService.StopGameDetection();
                 }
-                else if (settings.AutomaticGameDetection && !IsGameDetectionRunning)
+                else if (settings.AutomaticGameDetection && !_gameDetectionService.IsGameDetectionRunning)
                 {
-                    StartGameDetection();
+                    _gameDetectionService.StartGameDetection();
                 }
 
-                if (!settings.GameMemoryCommunication && IsGameCommunicationRunning)
+                if (!settings.GameMemoryCommunication && _gameCommunicationService.IsGameCommunicationRunning)
                 {
-                    StopGameCommunication();
+                    _gameCommunicationService.StopGameCommunication();
                 }
-                else if (settings.GameMemoryCommunication && !IsGameCommunicationRunning)
+                else if (settings.GameMemoryCommunication && !_gameCommunicationService.IsGameCommunicationRunning
+                    && _gameDetectionService.DetectedGame is not null)
                 {
-                    StartGameCommunication();
+                    _gameCommunicationService.StartGameCommunication(_gameDetectionService.DetectedGame.Process);
                 }
             });
         }
 
-        #region Game Communication 
-
-        private const int GAME_MEMORY_READ_INTERVAL = 1000;
-
-        private readonly SemaphoreSlim _memorySemaphore = new(1, 1);
-        private CancellationTokenSource _gameCommunicationCancellation = new();
-        private Task? _gameCommunicationTask;
-        private GameMemory? _gameMemory;
-
-        [MemberNotNullWhen(true, nameof(_gameCommunicationTask))]
-        public bool IsGameCommunicationRunning => _gameCommunicationTask != null && !_gameCommunicationTask.IsCompleted;
-        public GameState CurrentGameState { get; private set; } = new(false, default, null, null);
-
-        public event Action<GameState>? GameStateChanged;
-
-        public void StartGameCommunication()
-        {
-            if (IsGameCommunicationRunning ||
-                DetectedGame is null || DetectedGame.Process.HasExited)
-            {
-                _logger.LogDebug("Cannot start game memory communication: already running or no game process available");
-                return;
-            }
-
-            _memorySemaphore.Wait();
-            try
-            {
-                if (IsGameCommunicationRunning)
-                {
-                    return;
-                }
-
-                _logger.LogDebug("Starting game memory communication with {processName} ({pid})...",
-                    DetectedGame.Process.ProcessName, DetectedGame.Process.Id);
-
-                _gameMemory = new GameMemory(DetectedGame.Process, GAME_EXECUTABLE_NAME);
-                _gameCommunicationCancellation = new CancellationTokenSource();
-                _gameCommunicationTask = Task.Run(
-                    function: () => GameMemoryCommunicationLoop(_gameMemory, _gameCommunicationCancellation.Token)
-                                        .ContinueWith(t =>
-                                        {
-                                            if (t.IsFaulted)
-                                            {
-                                                _errorHandlingService.HandleError("Error during game communication");
-                                                _logger.LogError(t.Exception, "Game memory communication loop terminated with error:");
-                                            }
-                                            else if (t.IsCanceled)
-                                            {
-                                                _logger.LogInformation("Game memory communication loop canceled.");
-                                            }
-                                            else
-                                            {
-                                                _logger.LogInformation("Game memory communication loop terminated.");
-                                            }
-                                        }),
-                    cancellationToken: _gameDetectionCancellation.Token
-                 );
-
-
-                _logger.LogInformation("Game memory communication started with {processName} ({pid})",
-                    DetectedGame.Process.ProcessName, DetectedGame.Process.Id);
-            }
-            finally
-            {
-                _memorySemaphore.Release();
-            }
-        }
-
-        public void StopGameCommunication()
-        {
-            if (!IsGameCommunicationRunning)
-            {
-                return;
-            }
-
-            _logger.LogDebug("Stopping game memory communication...");
-
-            try
-            {
-                _gameCommunicationCancellation.Cancel();
-                _gameCommunicationTask.Wait();
-                _gameCommunicationTask = null;
-                _gameMemory?.Dispose();
-                _gameMemory = null;
-                CurrentGameState = new(false, ConnectionState.CA_DISCONNECTED, null, null);
-
-                _logger.LogInformation("Game memory communication stopped");
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error while stopping game communication");
-            }
-        }
-
-        private async Task GameMemoryCommunicationLoop(GameMemory gameMemory, CancellationToken cancellationToken)
-        {
-            while (!gameMemory.Process.HasExited && !cancellationToken.IsCancellationRequested)
-            {
-                ConnectState? connectState;
-                IPEndPoint? endPoint;
-
-                await _memorySemaphore.WaitAsync(cancellationToken);
-                try
-                {
-                    _logger.LogTrace("Reading memory from process {processName} ({pid})...",
-                        gameMemory.Process.ProcessName, gameMemory.Process.Id);
-
-                    ConnectionState connectionState = gameMemory.GetConnectionState() ?? ConnectionState.CA_DISCONNECTED;
-
-                    connectState = gameMemory.GetConnectState();
-                    if (connectState.HasValue)
-                    {
-                        try
-                        {
-                            IPAddress ipAddress = new(connectState.Value.Address.IP);
-                            int port = connectState.Value.Address.Port;
-                            endPoint = new IPEndPoint(ipAddress, port);
-                            _logger.LogTrace("Game connection state: {connectionState} - {endpoint}", connectionState, endPoint);
-                        }
-                        catch
-                        {
-                            endPoint = null;
-                        }
-                    }
-                    else
-                    {
-                        _logger.LogTrace("Game connection state: {connectionState}", connectionState);
-                        endPoint = null;
-                    }
-
-                    bool virtualLobbyLoaded = gameMemory.GetVirtualLobbyLoaded() ?? false;
-
-                    GameState lastGameState = CurrentGameState;
-                    if (virtualLobbyLoaded)
-                    {
-                        CurrentGameState = new(virtualLobbyLoaded, connectionState, null, null);
-                    }
-                    else
-                    {
-                        CurrentGameState = lastGameState with
-                        {
-                            ConnectionState = connectionState,
-                            VirtualLobbyLoaded = virtualLobbyLoaded,
-                            Endpoint = endPoint,
-                            StartTime = lastGameState.StartTime ?? DateTimeOffset.Now
-                        };
-                    }
-
-                    GameStateChanged?.Invoke(CurrentGameState);
-                    await Task.Delay(GAME_MEMORY_READ_INTERVAL, cancellationToken).ConfigureAwait(true);
-                }
-                finally
-                {
-                    _memorySemaphore.Release();
-                }
-            };
-        }
-
-        #endregion
-
-        #region Game Detection
-
-        private const int GAME_DETECTION_POLLING_INTERVAL = 1000;
-
-        private readonly object _gameDetectionLockObj = new();
-        private CancellationTokenSource _gameDetectionCancellation = new();
-        private Task? _gameDetectionTask;
-
-        public DetectedGame? DetectedGame { get; private set; }
-
-        [MemberNotNullWhen(true, nameof(_gameDetectionTask))]
-        public bool IsGameDetectionRunning => _gameDetectionTask != null && !_gameDetectionTask.IsCompleted;
-
-
-        public event Action<DetectedGame>? GameDetected;
-
-        public event Action? GameExited;
-
-        public void StartGameDetection()
-        {
-            if (IsGameDetectionRunning)
-            {
-                return;
-            }
-
-            lock (_gameDetectionLockObj)
-            {
-                if (IsGameDetectionRunning)
-                {
-                    return;
-                }
-
-                _logger.LogDebug("Starting game detection...");
-
-                _gameDetectionCancellation = new();
-                _gameDetectionTask = Task.Run(
-                    function: () => GameDetectionLoop(OnGameDetected, OnGameExited, cancellationToken: _gameDetectionCancellation.Token)
-                                        .ContinueWith(t =>
-                                        {
-                                            if (t.IsFaulted)
-                                            {
-                                                _errorHandlingService.HandleError("Game detection crashed");
-                                                _logger.LogError(t.Exception, "Game detection loop terminated with error:");
-                                            }
-                                            else if (t.IsCanceled)
-                                            {
-                                                _logger.LogInformation("Game detection loop canceled.");
-                                            }
-                                            else
-                                            {
-                                                _logger.LogInformation("Game detection loop terminated.");
-                                            }
-                                        }),
-                    cancellationToken: _gameDetectionCancellation.Token
-                 );
-
-                _logger.LogDebug("Game detection started");
-            }
-        }
-
-        public void StopGameDetection()
-        {
-            if (!IsGameDetectionRunning)
-            {
-                return;
-            }
-
-            _logger.LogDebug("Stopping game detection...");
-
-            lock (_gameDetectionLockObj)
-            {
-                try
-                {
-                    _gameDetectionCancellation.Cancel();
-                    _gameDetectionTask.Wait();
-                    _gameDetectionTask = null;
-
-                    _logger.LogInformation("Game detection stopped");
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error while stopping game detection");
-                }
-            }
-        }
-
-        private void OnGameDetected(DetectedGame detectedGame)
-        {
-            DetectedGame = detectedGame;
-
-            _logger.LogInformation("Detected game {gameProcessName} (v{gameVersion})",
-                detectedGame.Process.ProcessName, detectedGame.Version.ToString());
-
-            if (string.IsNullOrEmpty(_h2mLauncherSettings.CurrentValue.MWRLocation))
-            {
-                _logger.LogDebug("Game location empty, setting to {gameLocation}", detectedGame.FileName);
-
-                _h2mLauncherSettings.Update(settings =>
-                {
-                    return settings with
-                    {
-                        MWRLocation = detectedGame.FileName
-                    };
-                });
-            }
-
-            if (_h2mLauncherSettings.CurrentValue.GameMemoryCommunication)
-            {
-                StartGameCommunication();
-            }
-
-            GameDetected?.Invoke(detectedGame);
-        }
-
-        private void OnGameExited()
-        {
-            _logger.LogInformation("Game process exited");
-
-            DetectedGame = null;
-            if (IsGameCommunicationRunning)
-            {
-                StopGameCommunication();
-            }
-            GameExited?.Invoke();
-        }
-
-        private static async Task GameDetectionLoop(
-            Action<DetectedGame> onGameDetected, Action onGameExited, bool detectOnce = false, CancellationToken cancellationToken = default)
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                Process? process = FindH2MModProcess();
-                if (process is not null && process.MainModule is not null)
-                {
-                    string fileName = process.MainModule.FileName;
-                    string? gameDir = Path.GetDirectoryName(fileName);
-
-                    if (!string.IsNullOrEmpty(gameDir) && File.Exists(Path.Combine(gameDir, GAME_EXECUTABLE_NAME)))
-                    {
-                        FileVersionInfo version = FileVersionInfo.GetVersionInfo(fileName);
-
-                        // game dir found
-                        onGameDetected(new DetectedGame(process, fileName, gameDir, version));
-
-                        try
-                        {
-                            await process.WaitForExitAsync(cancellationToken).ConfigureAwait(false);
-                        }
-                        catch (InvalidOperationException)
-                        {
-                            // sometimes throws when proess is killed
-                        }
-
-                        // process terminated
-                        onGameExited();
-
-                        if (detectOnce)
-                        {
-                            return;
-                        }
-                    }
-                }
-
-                await Task.Delay(GAME_DETECTION_POLLING_INTERVAL, cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        #endregion
+        public IGameCommunicationService GameCommunication => _gameCommunicationService;
+        public IGameDetectionService GameDetection => _gameDetectionService;
 
         //Windows API functions to send input to a window
         [DllImport("user32.dll", CharSet = CharSet.Auto)]
@@ -405,18 +83,18 @@ namespace H2MLauncher.Core.Services
         private static extern bool EnumWindows(EnumWindowsProc enumProc, IntPtr lParam);
 
         [DllImport("user32.dll")]
-        internal static extern bool EnumThreadWindows(int dwThreadId, EnumWindowsProc lpfn, IntPtr lParam);
+        private static extern bool EnumThreadWindows(int dwThreadId, EnumWindowsProc lpfn, IntPtr lParam);
 
         [DllImport("user32.dll")]
-        internal static extern int GetWindowText(IntPtr hWnd, StringBuilder text, int count);
+        private static extern int GetWindowText(IntPtr hWnd, StringBuilder text, int count);
 
         [DllImport("user32.dll", SetLastError = true)]
-        static extern uint GetWindowThreadProcessId(IntPtr hWnd, out uint processId);
+        private static extern uint GetWindowThreadProcessId(IntPtr hWnd, out uint processId);
 
 
-        internal delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
+        private delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
 
-        internal static IEnumerable<IntPtr> EnumerateProcessWindowHandles(int processId)
+        private static IEnumerable<IntPtr> EnumerateProcessWindowHandles(int processId)
         {
             var handles = new List<IntPtr>();
 
@@ -427,7 +105,7 @@ namespace H2MLauncher.Core.Services
             return handles;
         }
 
-        internal static IEnumerable<IntPtr> EnumerateWindowHandles()
+        private static IEnumerable<IntPtr> EnumerateWindowHandles()
         {
             var handles = new List<IntPtr>();
 
@@ -624,7 +302,7 @@ namespace H2MLauncher.Core.Services
 
             // find process that loaded H1 MP binary
             var gameProc = processesWithTitle.FirstOrDefault(p =>
-                p.Modules.OfType<ProcessModule>().Any(m => m.ModuleName.Equals(GAME_EXECUTABLE_NAME)));
+                p.Modules.OfType<ProcessModule>().Any(m => m.ModuleName.Equals(Constants.GAME_EXECUTABLE_NAME)));
 
             return gameProc;
         }
@@ -632,7 +310,7 @@ namespace H2MLauncher.Core.Services
         private static bool IsH2MModProcess(Process p)
         {
             return p.MainWindowTitle.Contains("h2m-mod", StringComparison.OrdinalIgnoreCase) &&
-                p.Modules.OfType<ProcessModule>().Any(m => m.ModuleName.Equals(GAME_EXECUTABLE_NAME));
+                p.Modules.OfType<ProcessModule>().Any(m => m.ModuleName.Equals(Constants.GAME_EXECUTABLE_NAME));
         }
 
         private static IntPtr FindH2MModGameWindow(Process process)
@@ -653,18 +331,6 @@ namespace H2MLauncher.Core.Services
 
         public void Dispose()
         {
-            if (IsGameCommunicationRunning)
-            {
-                StopGameCommunication();
-            }
-
-            if (IsGameDetectionRunning)
-            {
-                StopGameDetection();
-            }
-
-            _gameMemory?.Dispose();
-            _memorySemaphore?.Dispose();
             _optionsChangeRegistration?.Dispose();
         }
     }
