+++ H2MLauncher.Core/Services/IGameDetectionService.cs
@@ -1,0 +1,13 @@
+﻿namespace H2MLauncher.Core.Services;
+
+public interface IGameDetectionService
+{
+    DetectedGame? DetectedGame { get; }
+    bool IsGameDetectionRunning { get; }
+
+    event Action<DetectedGame>? GameDetected;
+    event Action? GameExited;
+
+    void StartGameDetection();
+    void StopGameDetection();
+}
