+++ H2MLauncher.Core/Utilities/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace H2MLauncher.Core.Utilities
+{
+    internal class Constants
+    {
+        public const string GAME_EXECUTABLE_NAME = "h1_mp64_ship.exe";
+    }
+}
